requirements_termux.txt
slack-sdk
python-dotenv
